* kvalev/kvv-exif-face-regions:
  People: Update face region parsing and normalization documentation
  People: Improve debug logs
  People: Handle face regions in rotated images
  People: Additional MWG face region validation and logging
  People: Skip already indexed face regions
  People: Improve logging for exif face regions
  People: Create faces, subjects and markers from exif face regions
  People: Improve logging then parsing exif face regions
  People: Unique exif face regions
  People: Read IPTC face region metadata from exif
  People: Read MWG and WLPG face region metadata from exif

+++ internal/meta/data.go
@@ -99,19 +99,23 @@ type Region struct {
 }
 
 // Area describes the face area in MWG format.
+// All coordinates are relative to the image's orientation.
 type Area struct {
-	H    float32
-	W    float32
-	X    float32 // face center
-	Y    float32 // face center
+	H    float64 // face height
+	W    float64 // face width
+	X    float64 // face center
+	Y    float64 // face center
 	Unit string
 }
 
+// RegionIPTC represents a face region in IPTC format.
 type RegionIPTC struct {
 	Person   []string `json:"PersonInImage"`
 	Boundary Boundary `json:"RegionBoundary"`
 }
 
+// Boundary describes an IPTC face region shape.
+// All coordinates are relative to the image's orientation.
 type Boundary struct {
 	Shape string  `json:"RbShape"`
 	Unit  string  `json:"RbUnit"`
@@ -123,6 +127,7 @@ type Boundary struct {
 }
 
 // RegionMP represents a face region in WLPG format.
+// All coordinates are relative to the image's orientation.
 type RegionMP struct {
 	PersonDisplayName string
 	Rectangle         string
